requirements.txt
astroid==2.9.0
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cycler==0.11.0
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
fonttools==4.28.2
greenlet==1.1.2
idna==3.3
ImageHash==4.2.1
imageio==2.13.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.5.0
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.3
NudeNet==2.0.9
numpy==1.21.4
onnxruntime==1.9.0
opencv-python-headless==4.5.4.60
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
progressbar2==3.55.0
protobuf==3.19.1
pydload==1.0.9
pylint==2.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-utils==2.5.6
PyWavelets==1.2.0
regex==2021.11.10
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==1.4.27
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
